File created for each composables
Created a File for each composable in the mainActivity and added Preview for each of files.
